+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { NodeCreationModal } from "./node-creation-modal";
 export { NexusPropertiesSettingsTab } from "./settings/settings-tab";
+export { TitlePropertySetupModal } from "./title-property-setup-modal";